# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mobile Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List mobile devices
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sold
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Devices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    patch:
      operationId: updateDevice
      tags: [devices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

  # ─── CUSTOMERS ─────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer with purchase history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUPPLIERS ────────────────────────────────────────────────
...
  /bank-accounts:
    get:
      operationId: listBankAccounts
      tags: [bank-accounts]
      responses:
        "200":
          description: Bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank-accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}/transactions:
    get:
      operationId: listBankTransactions
      tags: [bank-accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bank transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankTransaction"
    post:
      operationId: createBankTransaction
      tags: [bank-accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"

  # ─── FAWRY / RECHARGE ─────────────────────────────────────────
...
components:
  schemas:
...
    Device:
      type: object
      required: [id, brand, model, imei1, condition, purchasePrice, purchaseDate]
      properties:
        id:
          type: integer
        brand:
          type: string
        model:
          type: string
        color:
          type: ["string", "null"]
        storage:
          type: ["string", "null"]
        imei1:
          type: string
        imei2:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
          enum: [new, used]
        purchaseDate:
          type: string
        purchasePrice:
          type: number
        salePrice:
          type: ["number", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        sold:
          type: boolean
        soldAt:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
    DeviceInput:
      type: object
      required: [brand, model, imei1, condition, purchasePrice, purchaseDate]
      properties:
        brand:
          type: string
        model:
          type: string
        color:
          type: string
        storage:
          type: string
        imei1:
          type: string
        imei2:
          type: string
        serialNumber:
          type: string
        condition:
          type: string
          enum: [new, used]
        purchaseDate:
          type: string
        purchasePrice:
          type: number
        salePrice:
          type: number
        supplierId:
          type: integer
    DeviceUpdate:
      type: object
      properties:
        color:
          type: string
        storage:
          type: string
        salePrice:
          type: number
        sold:
          type: boolean
        soldAt:
          type: string
        customerId:
          type: integer

    # ─── CUSTOMERS ──────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalDebt:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        nationalId:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        nationalId:
          type: string
        notes:
          type: string
    CustomerDetail:
      type: object
      required: [id, name, phone, totalDebt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalDebt:
          type: number
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        installments:
          type: array
          items:
            $ref: "#/components/schemas/InstallmentPlan"
        maintenanceOrders:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceOrder"

    # ─── SUPPLIERS ──────────────────────────────────────────────
...
    Sale:
      type: object
      required: [id, invoiceNumber, customerId, customerName, subtotal, discount, total, paymentMethod, paidAmount, dueAmount, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, etisalat_cash, we_pay, instapay, bank_transfer]
        paidAmount:
          type: number
        dueAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MaintenanceOrder:
      type: object
      required: [id, ticketNumber, customerName, customerPhone, deviceType, issue, status, createdAt]
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        deviceType:
          type: string
        deviceBrand:
          type: ["string", "null"]
        issue:
          type: string
        password:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, inspecting, repairing, waiting_parts, repaired, delivered]
        estimatedCost:
          type: ["number", "null"]
        finalCost:
          type: ["number", "null"]
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    InstallmentPlan:
      type: object
      required: [id, customerId, customerName, customerPhone, deviceName, totalAmount, downPayment, installmentAmount, totalInstallments, paidInstallments, remainingAmount, nextDueDate, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deviceName:
          type: string
        totalAmount:
          type: number
        downPayment:
          type: number
        installmentAmount:
          type: number
        totalInstallments:
          type: integer
        paidInstallments:
          type: integer
        remainingAmount:
          type: number
        nextDueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, overdue]
        createdAt:
          type: string
...
    BankAccount:
      type: object
      required: [id, bankName, accountNumber, balance]
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountNumber:
          type: string
        iban:
          type: ["string", "null"]
        instapayNumber:
          type: ["string", "null"]
        balance:
          type: number
    BankAccountInput:
      type: object
      required: [bankName, accountNumber]
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        iban:
          type: string
        instapayNumber:
          type: string
        balance:
          type: number
    BankTransaction:
      type: object
      required: [id, bankAccountId, type, amount, balanceBefore, balanceAfter, createdAt]
      properties:
        id:
          type: integer
        bankAccountId:
          type: integer
        type:
          type: string
          enum: [deposit, withdraw, transfer]
        amount:
          type: number
        fee:
          type: number
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BankTransactionInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [deposit, withdraw, transfer]
        amount:
          type: number
        fee:
          type: number
        notes:
          type: string

    # ─── FAWRY ───────────────────────────────────────────────────
...